refactor(nc-gui): move panels into separate components
