+++ packages/nc-gui/components/erd/HistogramPanel.vue
@@ -1,0 +1,19 @@
+<script lang="ts" setup>
+import { Panel } from '@vue-flow/additional-components'
+</script>
+
+<template>
+  <Panel class="text-xs bg-white border-1 rounded border-gray-200 z-50 nc-erd-histogram" position="bottom-right">
+    <div class="flex flex-col divide-y-1">
+      <div class="flex items-center gap-1 p-2">
+        <MdiTableLarge class="text-primary" />
+        <div>{{ $t('objects.table') }}</div>
+      </div>
+
+      <div class="flex items-center gap-1 p-2">
+        <MdiEyeCircleOutline class="text-primary" />
+        <div>{{ $t('objects.sqlVIew') }}</div>
+      </div>
+    </div>
+  </Panel>
+</template>
